Create README.md
SGF Freight had a website developed in 2007.  This site was a static website and most of the links were broken and none functional, so, I had to redesign the site layout (Wireframe) recoded the html from the scratch, added some features with js and jquery and made the site responsive.